Don't assert on empty directories.
When run on a directory with no cts data at all, the report
generator for `hooktest --console` would throw a StatisticsError
trying to calculate the mean coverage over zero files. Check
for this condition and report zero coverage instead.

Add a test for this condition, invoking hooktest on the included
emptydir fixture.